catkin_ws/src/ros_deeplabv3/requirements.txt
rospy
numpy
torch
torchvision
Pillow
pyyaml
opencv-python